[tricium] Fix the post comment requests to use commit message file
In the post to Gerrit, the request body includes a map of path to comments,
and the comments themselves which also include the path. Previously
the magic path for commit message was used in the comment but not in
the map.

This CL also extracts a small function to avoid repetitive code.

Bug: 961422
Change-Id: If1db4239545305da7fd5b50816150b0a5cdf07fd
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1680716
Reviewed-by: Marc-Antoine Ruel <maruel@chromium.org>
Commit-Queue: Marc-Antoine Ruel <maruel@chromium.org>
Cr-Commit-Position: refs/heads/master@{#24003}