[buildbucket] Parse output commit from legacy
This CL is an emergency measure for crbug.com/940214 which blocks
crbug.com/970053.
A better, but much more risky and involved way of doing this is calling
api.buildbucket.set_output_gitiles_commit in recipes.

Note that the hack is being added to code that is scheduled for deletion.

R=iannucci@chromium.org, tandrii@chromium.org

Bug: 940214
Change-Id: I24ab70641d11cafd0759204290aeec99c951421b
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1682204
Auto-Submit: Nodir Turakulov <nodir@chromium.org>
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Cr-Commit-Position: refs/heads/master@{#24019}