cros_test_platform: add scaffolding for isolate client
The skylab-execute step will need to pull task results via an isolate
client. This CL adds the scaffolding for it to do so, in the form of a
Client interface.

In order to keep parity between the autotest and skylab runners, the
client interface is also added to the autotest runner, although it will
not be used.

BUG=971376
TEST=None

Change-Id: I2d1dce472204da26db5056120c9a1e2a39d995cc
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1682115
Commit-Queue: Aviv Keshet <akeshet@chromium.org>
Reviewed-by: Alex Zamorzaev <zamorzaev@chromium.org>
Cr-Commit-Position: refs/heads/master@{#24032}