[tricium] Add a fixed offset to line numbers in the commit message
Tricium can now pass the commit message to analyzers and posts
resulting comments on the CL, but the line numbers must be adjusted
because Gerrit adds extra header lines to the commit message "file".

The commit message that Gerrit provides is always without headers,
so there's no clear way to get the number of header lines from
Gerrit... but the number of header lines always appears to be the
same, so adding a fixed offset to the line numbers on comments on
the commit message will probably work, and I don't see another
obvious solution here.

Bug: 961422
Change-Id: I0586eaac8e0eb23fb2ccbf43c2e9eedaa15aaeb4
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1684135
Commit-Queue: Quinten Yearsley <qyearsley@chromium.org>
Reviewed-by: Andrii Shyshkalov <tandrii@chromium.org>
Cr-Commit-Position: refs/heads/master@{#24075}