generalize DeployDut -- multiple new-specs
Change the new_specs field of the DeployDut protobuf definition to be
a list of DUT descriptions rather than a single DUT description.

This is a backwards incompatible change to the DeployDut API, but it's
safe because DeployDut has no callers at the moment.

Bug: 976324
Change-Id: Iece564a69449d6622a88de528cade906121da155
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1670168
Commit-Queue: Gregory Nisbet <gregorynisbet@google.com>
Reviewed-by: Allen Li <ayatane@chromium.org>
Reviewed-by: Xixuan Wu <xixuan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#24082}