[tricium] Remove the Url field from Data_Comment
This field is not actually used. It was added because Gerrit
RobotCommentInfo has a URL field -- but the problem is that
the text of the link cannot be controlled and the text is always
the robot ID (for Tricium we use the run ID).

The fact that the field is there adds confusion because it
looks like it would be used but it's not.

Change-Id: Icf1eb2c150085caae9dbf8889b4c92b6e1ef1cd4
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1682915
Commit-Queue: Quinten Yearsley <qyearsley@chromium.org>
Reviewed-by: Marc-Antoine Ruel <maruel@chromium.org>
Cr-Commit-Position: refs/heads/master@{#24093}