[code coverage] Update coverage proto to use summaries for File
This CL updates the property name for metrics to be consistent with
GroupCoverageSummary and CoverageReport.

This mis-naming was not caught because the data was never filled in,
and there will be a follow-up CL to fill in the data, and the main use
case is that the summaries for the files will be used to calculate the
coverage summaries for dirs and components.

Bug: 980257
Change-Id: Ibcf246be5fb3c3857050231f1f4277295c8a4bc0
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1684147
Reviewed-by: Roberto Carrillo <robertocn@chromium.org>
Commit-Queue: Yuke Liao <liaoyuke@chromium.org>
Cr-Commit-Position: refs/heads/master@{#24109}