cros_test_platform: pull build name into autotest proxy request
Note: we are currently pulling the build name label out of the build
URL; that is incorrect. In the near future, we will need to either add a
URL->provisionable label translation library, or change the request
proto so that provisionable labels are specified as labels.

BUG=971377
TEST=None

Change-Id: I3249f0e30859284eb76361fa544b22e8fcef5f77
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1686430
Commit-Queue: Aviv Keshet <akeshet@chromium.org>
Reviewed-by: Alex Zamorzaev <zamorzaev@chromium.org>
Cr-Commit-Position: refs/heads/master@{#24145}