cros_test_platform: use a GetterFactory instead of Getter in executors
Prior to this CL, we relied on creating an isolate client ahead of time,
and then using it once tasks gave output refs. However, the isolate
server to use is actually specified by the output ref itself, and not
necessarily known at launch time. Use a GetterFactory instead so that
the isolate client can be created at the time when the server address is
known.

BUG=971376
TEST=None

Change-Id: I476fb3b7b1f6514e07f0190b86afc8402d3debeb
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1693344
Reviewed-by: Alex Zamorzaev <zamorzaev@chromium.org>
Commit-Queue: Aviv Keshet <akeshet@chromium.org>
Cr-Commit-Position: refs/heads/master@{#24255}