Use end_time for start_time if the given field set doesn't have a start time.
Logs show that that _TargetFields.start_times happens to miss an entry for
a given field key that are present in _TargetFields.values.

The cause of the missing key exception is not known yet. Looking at the
implementation of _TargetFields member methods, I haven't been able to find
a spot that could possibly cause such an error.

Further investigation is required to find the cause of the error, but this CL
is to stop the missing key exception being raised in cr-buildbucket.

Bug: 980625
Change-Id: Id07a57675b7b1d84d829388a3f91fe23e007936b
R: nodir,vadimsh
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1687152
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Reviewed-by: Nodir Turakulov <nodir@chromium.org>
Commit-Queue: Scott Lee <ddoman@chromium.org>
Cr-Commit-Position: refs/heads/master@{#24328}