crosskylabadmin: modify fake inventory data to check IsEmptyErr.
IsEmptyErr is not caught in DeleteDuts(). It should be found via test
TestDeleteDuts: DeleteDuts with unknown hostname deletes no duts. But it
doesn't.

It's because in test_helper, the fake inventory data doesn't have the
same format as real inventory data. So even if there's no real change
between old and new inventory, they're different when they're converted
to string. So when testing with removing unknown hostname, a change is
still made, and TestDeleteDuts won't raise error.

This CL changes the fake inventory data format to align to the format of
real inventory data, so that we can check if IsEmptyErr is properly
caught.

Change-Id: I0762732be98a91d1b474e7cdc9f6585a47f4a4ac
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1697179
Commit-Queue: Xixuan Wu <xixuan@chromium.org>
Auto-Submit: Xixuan Wu <xixuan@chromium.org>
Reviewed-by: Allen Li <ayatane@chromium.org>
Cr-Commit-Position: refs/heads/master@{#24334}