qscheduler: only update examinedTime on prioritized requests
This CL improves performance and fixes a subtle bug. We should not be
updating the examined time on requests that are non-throttled (from
fanout) but still unable to be matched due to belonging to a
free-tasks-disabled account that is out of quota.

BUG=982910
TEST=New unit tests added, and confirmed to fail prior to this CL.
benchmark time drops from ~9.2s to ~8s.

Change-Id: I447e1e273f16a63720a195fa46ef7eff4abe7949
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1699481
Commit-Queue: Aviv Keshet <akeshet@chromium.org>
Reviewed-by: Andrii Shyshkalov <tandrii@chromium.org>
Cr-Commit-Position: refs/heads/master@{#24422}