[buildbucket] Improve BQ export
Complain if we didn't insert a single row. This helps when bb server is
newer than BQ table schema. Without this, we don't notice a permanent
failure because of bb's retry-later mechanism.

Run export in a tight loop if there were >= 100 builds to insert.
This helps if there is a backlog (due to prior outage).
Without this, bb inserts 300 builds per minute and it is too slow to process
tens of thousands of backlogged builds on top of builds being inserted
constantly.

Change-Id: I322acef316c09782235a5dd26b7868745b6f7a26
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1699193
Commit-Queue: Nodir Turakulov <nodir@chromium.org>
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Cr-Commit-Position: refs/heads/master@{#24548}