[Monorail]  Add new TF2.0 spam and component ML models
This CL adds the ability to train new ML models locally for spam and component prediction. The models are written in TensorFlow 2.0, which is currently in beta. They're added to a new directory instead of replacing the old ones so that we can continue to train TensorFlow 1.x models until we no longer use them in production.

Bug: monorail:5898, monorail:5899
Change-Id: Ifffcd338bbe0f85d79dc6b320af7e49f15184609
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1707561
Commit-Queue: Anita Li <weihanl@google.com>
Reviewed-by: Jeff Carpenter <jeffcarp@chromium.org>
Cr-Commit-Position: refs/heads/master@{#24579}