[tricium spellchecker] Make the ignore patterns more inclusive
Specifically, this CL would make it so that strings that include
an email address or URL are ignored, even if there are extra
characters before or after the email address or URL, e.g.
"TBR=alph@chromium.org,...".

maruel@ suggested adding ^ to the start of the patterns earlier
because in general it often makes regexes clearer, although in
this case we may actually want to match the url/email patterns
anywhere even when there are other characters things before,
like in the above example.

Note: TODOs are a bit of a different case, because they're supposed
to be at the beginning of a line, often with a line comment token
before it, so I think that "...abc...TODO...xyz..." should perhaps
still be checked.

Bug: 985414
Change-Id: If806c3c3d0fa76c9d365bb184d648af4079999b8
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1709512
Commit-Queue: Quinten Yearsley <qyearsley@chromium.org>
Reviewed-by: Andrii Shyshkalov <tandrii@chromium.org>
Cr-Commit-Position: refs/heads/master@{#24607}