[tricium] Remove the FanOutIn when updating entities in worker_done
This is a preliminary change that I'd like to make before
adding extra comment filtering in createCommentSelections
(https://chromium-review.googlesource.com/c/infra/infra/+/1691076).

As far as I understand, the parallelization here is not necessary
and in theory the writes will be serialized anyway, so doing all of
these datastore operations in sequence won't be significantly slower.

Note-- There are a couple of other small changes I'd like to make
here as well, which I plan to do in separate CLs; the main one is that
createCommentSelections should be moved out of the RunInTransaction,
since it now is a relatively expensive function which does queries
across multiple entity groups.

Bug: 890545
Change-Id: I356bd890566f781e53079d41486a2dc9d7f0b6ae
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1717489
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Reviewed-by: smut <smut@google.com>
Commit-Queue: Quinten Yearsley <qyearsley@chromium.org>
Cr-Commit-Position: refs/heads/master@{#24635}