[Monorail] Remove param owner_email and labels from ProcessAlert().
This CL removes param owner_email because it's always to be None in the current
implementation. Creating an issue with the an owner will be supported by
the following email header speicifed in monitoring configs.
: http://shortn/_Qouo9ylYMv
However, the owner info will not be passed as an individual parameter to
ParseAlert(), anyways, because there are too many email headers to pass each
one as an individual parameter to the method.

This CL replaces param labels with label, because the regex supports only
one single label, and, therefore, param labels always contains one or 0 label.
: https://cs.chromium.org/chromium/infra/appengine/monorail/framework/emailfmt.py?l=250-253

Change-Id: I6ccad008c434f25fcb7d64c1a63be4b26cecfbff
Bug: 807064
R: zhangtiff
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1716689
Reviewed-by: Tiffany Zhang <zhangtiff@chromium.org>
Commit-Queue: Scott Lee <ddoman@chromium.org>
Cr-Commit-Position: refs/heads/master@{#24658}