[Monorail] Allow alert issues to be updated, regardless of the last update time.
The logic assumes that an alert email is a duplicate to an existing one,
if the existing one was posted in the last 24 hrs.

It's not a safe assumption, and this CL is to remove the dedup logic.
For more details about incident notifications, please find the description
of crbug.com/987551

BUG=807064,987551

Change-Id: I4792e3540aaffe545b58450579250121bdc76c20
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1716691
Reviewed-by: Tiffany Zhang <zhangtiff@chromium.org>
Commit-Queue: Scott Lee <ddoman@chromium.org>
Cr-Commit-Position: refs/heads/master@{#24659}