[tricium] Suppress comments if previously marked "not useful"
In this CL: In the worker_done handler, around the same
time that comments outside of the changed lines are filtered,
also fetch past comments that were marked as not useful,
and filter out any new comments similar to the past comments.

This change also requires changing the properties GerritHost
and GerritChange in AnalyzeRequest to be indexed, and adds
another required index in index.yaml.

The overall purpose of this change is to reduce the amount of
noise and "annoyance factor" of repeated false positives.

Bug: 890545
Change-Id: If29694163e20f922b4be2f3b840a28d0b1e9348b
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1691076
Commit-Queue: Quinten Yearsley <qyearsley@chromium.org>
Reviewed-by: smut <smut@google.com>
Cr-Commit-Position: refs/heads/master@{#24681}