[cloudbuildhelper] Tag docker images.
Introduces a notion of "canonical tag", which is "immutable" (by convention)
docker tag for a reproducibly built images. Before building and pushing
<image>:<canonical-tag> we check whether there's such image in the registry
already and completely skip the build if it is so.

We'll use (perhaps somehow formatted) git revisions of infra.git as our canonical
tags.

R=iannucci@chromium.org
BUG=985561

Change-Id: I2bd80e4021eff4ce942d036b3134b913fb42f1b2
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1722210
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Cr-Commit-Position: refs/heads/master@{#24714}