[Monorail] Add CreateAlertIssueProperties().
The method is expected to contain all the logic parsing and merging all
the information passed via the email-address and email-headers.

In this CL, the method just contains all the existing logic handling
the information passed via the email-address, but, in a follow-up CL, more
logic will be added to handle email headers with additional unit tests.

Bug: 807064
Change-Id: I2c16efcd62e7ba2aa2514572bd5de9c9de9ee731
R: zhangtiff
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1722911
Reviewed-by: Tiffany Zhang <zhangtiff@chromium.org>
Commit-Queue: Scott Lee <ddoman@chromium.org>
Cr-Commit-Position: refs/heads/master@{#24721}