mitigate addTaskInfoSummaries timeout
This is a VERY temporary fix. It's not intended to be a long term solution to the
Refresh-Bots timeout problem.

The refresh bots task is timing out.
This is a serious problem.
Right now as far as I can tell the only symptoms are that `skylab diagnose` and
`skylab inventory` no longer work.

This CL is a hacky potential mitigation. It randomly selects half of the BOTs
and updates those.

I chose `getBotsFromSwarming` as the choke point, since it already post-processes the
response from the swarming client before handing said response to the rest of the
application.

Because we have evidence from the logs that control is transferred to `addTaskInfoSummaries`,
I am confident that `getBotsFromSwarming` is not itself the thing that is timing out.

Bug: chromium:989980
Change-Id: I91ccdeae020fcb359b8ff91041c4d2490ae8ca9c
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1731695
Reviewed-by: Allen Li <ayatane@chromium.org>
Commit-Queue: Gregory Nisbet <gregorynisbet@google.com>
Cr-Commit-Position: refs/heads/master@{#24835}