Add unit type to few metrics in cros sysmon
Added unit types to memory, time and temperature related metrics in cros
sysmon. Memory is in 'KiB' because /proc/meminfo outputs as 'kB' but
that is due to historic reasons and the actual value is in Kibibytes.
(https://lore.kernel.org/patchwork/patch/666444/)
Temperature is in °C, DUT outputs temperature in int units of 10^(-3)°C
values, which is converted to float units of °C

R=bpastene@google.com

Bug: 937943
Change-Id: I7452eb7b880670e4b87648b969b3162b424c7665
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1742647
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Reviewed-by: Scott Lee <ddoman@chromium.org>
Commit-Queue: Anushruth  <anushruth@chromium.org>
Cr-Commit-Position: refs/heads/master@{#24976}