[code coverage] Correct list of files to rebase between patchsets
Originally, the list of files to rebase is based on the files changed
by the src patchset, however, this is not correct because two patchsets
could have different list of changed files even though they're trivial
rebase away.

This CL fixes the issue by querying Gerrit to fetch the list of files
added and modified by the dest patchset and use it as the files to
rebase.

Bug: 910289
Change-Id: I794e7f0c79ff9dc61cf4b5abd63489c43a5ee28e
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1735051
Commit-Queue: Yuke Liao <liaoyuke@chromium.org>
Reviewed-by: Sajjad Mirza <sajjadm@chromium.org>
Reviewed-by: Nodir Turakulov <nodir@chromium.org>
Cr-Commit-Position: refs/heads/master@{#25021}