qscheduler: stop using Config proto in scheduler logic
This CL removes direct references from scheduler logic to the Config
proto, instead wrapping the Config in a native struct (much the same to
how scheduler.State is a wrapping the State proto). This allows for
richer type information in the native struct, and better type safety.

BUG=992713
TEST=None

Change-Id: Ibbc0a055dd9212822eeba3de8af09e1821954c76
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1749733
Commit-Queue: Aviv Keshet <akeshet@chromium.org>
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Cr-Commit-Position: refs/heads/master@{#25034}