Add partial matching functionality to step_util
The step_name retrieved from the test_results table is often a prefix of
how the full step name appears in the build, therefore it is not
currently possible to retrieve step metadata using GetStepMetadata.

Solution:
Added a parameter partial_matching which defaults to False to functions
in step_util that are used in retrieving step metadata.

Currently, _GetStepLogViewUrl only returns a step log url if the
step_name is completely matches a step as listed in the builder. In the
case that the step_name does not completely match a result, the
partial_matching flag allows _GetStepLogViewUrl to return a log url if
step_name is a prefix of a step in the builder.

Other: Changed the name of GetPlatform to GetOS as suggested in
CL/1652534.

Bug: 992113
Change-Id: I8a41460fa3cfe32bb42bae460660c088259f61f9
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1745520
Commit-Queue: Navil Perez <navil@google.com>
Reviewed-by: Yuke Liao <liaoyuke@chromium.org>
Reviewed-by: Chan Li <chanli@chromium.org>
Cr-Commit-Position: refs/heads/master@{#25035}