[dockerbuild] make --wheel_re a little more user friendly
It's not obvious whether --wheel_re allows partial matches (it
doesn't), and if it doesn't match anything, we fallback to building
the default set.  Change the behavior to exit immediately if the
user's regex didn't match anything, and include the full regex the
code used so people can debug further.

Bug: 866772
Test: `./run.py infra.tools.dockerbuild wheel-build --wheel_re 'pylint' --rebuild` aborts
Test: `./run.py infra.tools.dockerbuild wheel-build --wheel_re 'pylint.*' --rebuild` builds pylint
Change-Id: I6f05a93b1b3ecdb37d14ff33709cdb0ed05c2934
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1754405
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>
Cr-Commit-Position: refs/heads/master@{#25099}