CrosSkylabAdmin: Deploy multiple duts simultaneously.
If there are no actions such as installing the OS
(what is the generic term for this type of deployment action?) then
deploy multiple DUTs in the same gerrit commit.

I'm trying to support the `skylab quick-add-duts` command line thing.
Right now the implementation in CrOSSkylabAdmin backing `skylab quick-add-duts`
generates a single commit per DUT, which is not scalable.

If there are actions to perform to deploy a DUT after adding it to the inventory,
then that will be done the same way in the world post-[this diff] as it is in
the status quo.

Bug: chromium:991640
Change-Id: Ia24d5331f064918e69a91ceffb8ea96194ee7040
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1742877
Reviewed-by: Allen Li <ayatane@chromium.org>
Commit-Queue: Gregory Nisbet <gregorynisbet@google.com>
Auto-Submit: Gregory Nisbet <gregorynisbet@google.com>
Cr-Commit-Position: refs/heads/master@{#25168}