[Monorail] Load gapi.js for federated references
This CL introduces the plumbing to load gapi.js for future use fetching
data from Buganizer. There are no user-facing changes in this CL. If a
user navigates to a page with a federated reference on it, it will load
gapi.js but do nothing else.

Bug: monorail:5939
Change-Id: If6e0a433993512572b494fff97d004e4ea0e303a
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1704795
Commit-Queue: Jeff Carpenter <jeffcarp@chromium.org>
Reviewed-by: Tiffany Zhang <zhangtiff@chromium.org>
Cr-Commit-Position: refs/heads/master@{#25262}