[code coverage] Implement report timestamp metric.
With this metric it will be easier to query how long it has been since a
report was generated, and from it we can derive the uptime of the
service.

After a transition period we'll deprecate the cumulative full-reports
metric, as the new one should be able to replace it in alerts and
dashboards.

R=liaoyuke,nodir

Bug: 992127
Change-Id: I16e4eb49a1882dbeac2837f509cd26db1a938829
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1761414
Reviewed-by: Roberto Carrillo <robertocn@chromium.org>
Reviewed-by: Yuke Liao <liaoyuke@chromium.org>
Commit-Queue: Roberto Carrillo <robertocn@chromium.org>
Cr-Commit-Position: refs/heads/master@{#25375}