[led] Set logdog prefix as led_run_id property
Passes the logdog prefix as a led recipe module property so that led
jobs can use it as a unique identifier (specifically, as a GCS path for
uploading build artifacts to).

This required moving the generation of the logdog prefix from job.go
up a level into cmd_launch.go.

Bug: fuchsia:35999
Change-Id: Ia1ea63ef0f4d884dd0211027b911b9b37378e28b
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1783704
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Commit-Queue: Oliver Newman <olivernewman@google.com>
Cr-Commit-Position: refs/heads/master@{#25592}