[Monorail] Fix Updates nav link when viewing own profile.
One reason why the user may not have discovered this page is that the
nav link to it goes to a page that we are phasing out.  The user menu
link is correct, but if the user navigated to their own profile and
then used a top-of-page tab, it went to updates in starred projects.
That page is being phased out to simplify the UI and because it had
low usage.  Although there are thousands of project stars in our DB,
very few of them are from Monorail, most are from code.google.com.

R=ehmaldonado@chromium.org

Bug: monorail:6308
Change-Id: I7b2c144350509de529ad04005736175c68304d56
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1790090
Reviewed-by: Edward Lesmes <ehmaldonado@chromium.org>
Commit-Queue: Jason Robbins <jrobbins@chromium.org>
Cr-Commit-Position: refs/heads/master@{#25612}