Adds associated bugs to a LuciTest entity.
Adds an issue_key attribute to LuciTest in order to hold FlakeIssue keys
for associated bugs.

Pipeline to add issue_keys to LuciTest for new bugs found in the
disabled tests query. If a FlakeIssue does not already exist for a bug,
it creates one.

Bug: 1001130
Change-Id: I2845a6af1513f5ef5619f1814f722b40b193022d
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1787236
Reviewed-by: Yuke Liao <liaoyuke@chromium.org>
Reviewed-by: Chan Li <chanli@chromium.org>
Commit-Queue: Navil Perez <navil@google.com>
Cr-Commit-Position: refs/heads/master@{#25613}