[Monorail] Redirect from bugs.fuchsia.dev homepage to project.
This adds logic to the HostingHome servlet to make it redirect if the
request hostport matches a value configured in settings.py.

To test in staging, the new version must be made live in staging.  Then,
access it via the branded domain bugs-staging.fuchsia.dev using an
account that either can or cannot view the fuchsia project.

R=jojwang@chromium.org

Bug: monorail:6270
Change-Id: I21ff5a4217e8608305cc086cda8665fdff21ab01
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1790093
Commit-Queue: Jason Robbins <jrobbins@chromium.org>
Reviewed-by: Joanna Wang <jojwang@chromium.org>
Cr-Commit-Position: refs/heads/master@{#25616}