[code coverage] Fix a bug that total_lines in CoveragePercentage is 1
There is a bug in the code that makes all the total_lines in
CoveragePercentage to be 1.

The root cause is that I mis-understood how validator works, it's
supposed to return a validated/modified value or raise an exception,
instead of returning a boolean value.

Bug: 939904
Change-Id: I939f3e08c0caa3b9897b68d516dd70135b62c8b5
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1795407
Reviewed-by: Roberto Carrillo <robertocn@chromium.org>
Commit-Queue: Yuke Liao <liaoyuke@chromium.org>
Cr-Commit-Position: refs/heads/master@{#25654}