Add missing test data for output placeholders
In preparation for this CL:
https://chromium-review.googlesource.com/c/infra/luci/recipes-py/+/1777143

Currently, if no test step data is provided for a step that uses an
output placeholder, the placeholder evaluates to an empty string. After
that CL lands, the placeholder will instead evaluate to None, breaking
any non-mocked code that assumes a non-null placeholder result.

Change-Id: I5a1523ca1bb219abd4fa4c0d3b782f90b5052d34
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1796995
Commit-Queue: Oliver Newman <olivernewman@google.com>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Cr-Commit-Position: refs/heads/master@{#25671}