[tricium] Remove incomplete Fix parsing
The previous CL added a string field for the new "fix" field in
shellcheck 0.7 output.  The field is actually a nested object with a
complicated structure.  We don't yet have a good way to expose the
suggested fixes in tricium, so let's just remove the field for now
instead of fully parsing it.  Added new tests cases that trigger the fix
behavior to verify that it is skipped correctly.

Bug: 989299
Test: make testrun
Change-Id: I41f5d8be7756fc48c13f981cca49a3f977c75d62
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1796591
Commit-Queue: Benjamin Gordon <bmgordon@chromium.org>
Reviewed-by: Quinten Yearsley <qyearsley@chromium.org>
Cr-Commit-Position: refs/heads/master@{#25712}