Hide disabled tests from Flake Portal
Hide flakes from flake portal:
A test will be hidden from flake portal if there exists a LuciTest
entity for it that is stored as disabled.

Also in this change, do not update monoral bugs for a test if the
FlakeIssue is being tracked by a LuciTest entity.

Bug: 953515, 995011
Change-Id: I5af85c1a5feaca298c2dd4ae02ae4beacdb9ac55
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1797268
Commit-Queue: Navil Perez <navil@google.com>
Reviewed-by: Chan Li <chanli@chromium.org>
Cr-Commit-Position: refs/heads/master@{#25728}