Add package-lock.json to source control
This file is generated when running "make build", it is recommended that it should be committed. I also checked with SoM and other services and they did contain this file as well.

Also, I think if we leave this file uncommitted, the deployment will be tainted.

R=orodley@google.com

Bug: 1005557
Change-Id: Ief639842bee53f9382bde3f567008a178d394a84
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1810148
Reviewed-by: Owen Rodley <orodley@chromium.org>
Commit-Queue: Quang Minh Tuan Nguyen <nqmtuan@google.com>
Cr-Commit-Position: refs/heads/master@{#25863}