qscheduler: include all labels in fanout group determination
This CL changes the behavior of quotascheduler, so that the parallelism
limits it applies are enforced among tasks that are similar in all their
supplied labels, not merely in their supplied provisionable labels.

This is necessary because there are some ChromeOS clients that use the
same provisionable label (a given unibuild build) against a wide variety
of device types.

BUG=1005553
TEST=unit test updated; test verified to fail prior to fix. Scheduler
benchmark regresses from ~5.7s to 8.3s, but this is to be expected
because the simulation behavior is quite different.

Change-Id: I7831f413419e4f49e7dca9174b0ab160eea5cebf
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1812345
Commit-Queue: Aviv Keshet <akeshet@chromium.org>
Reviewed-by: Andrii Shyshkalov <tandrii@google.com>
Reviewed-by: C Shapiro <shapiroc@chromium.org>
Cr-Commit-Position: refs/heads/master@{#25894}