Linkify BlockedOn/Blocking/MergedInfo in list view
Issues are clickable as links.  If the issue is already resolved, it is
displayed with a line through it.

Testing:
Ran a local instance of monorail and added multiple bugs with blocking and
duplicates issues.  Then view them as a list with this link:
http://localhost:8081/p/monorail-testing/issues/list?colspec=ID%20Type%20Status%20Priority%20Milestone%20Owner%20Summary%20BlockedOn%20Blocking%20MergedInto&q=&can=1

Bug: monorail:6280
Change-Id: I6a382152e8bb299f0f055a4e6babaafff350f1ce
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1812029
Reviewed-by: Tiffany Zhang <zhangtiff@chromium.org>
Commit-Queue: Eyal Soha <eyalsoha@google.com>
Cr-Commit-Position: refs/heads/master@{#25903}