[Monorail] Fix mergedinto FedRef getting clipped in the UI
Because the surrounding <td> has the CSS rule `overflow: hidden` to enable
`text-overflow: ellipsis` to work properly, it cuts off the info dropdown for
federated references in the mergedinto field. This CL special cases that field,
with a TODO to find a more elegant solution.

Bug: 6369
Change-Id: Ia5e8f06c2f1a4f36acdb79834b11e123c44955b0
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1816835
Reviewed-by: Tiffany Zhang <zhangtiff@chromium.org>
Commit-Queue: Jeff Carpenter <jeffcarp@chromium.org>
Cr-Commit-Position: refs/heads/master@{#25994}