[Findit V2] Allow projects to provide dimensions for tryjobs.
This is needed by the chromium project, as it uses a single rerun
builder backed with slaves of different platforms.

This change makes the chromium project copy the 'os' and 'gpu'
dimensions of the build being analyzed to the requested dimensions of
the tryjobs used to analyze such failures.

R=chanli

Bug: 978503
Change-Id: I245e5bfbd2b00bfcbe1458bde0ffdc5cab540e5b
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1828282
Reviewed-by: Chan Li <chanli@chromium.org>
Commit-Queue: Roberto Carrillo <robertocn@chromium.org>
Cr-Commit-Position: refs/heads/master@{#26093}