Linkify BlockedOn/Blocking/MergedInfo in list view
Issues are clickable as links.  If the issue is already resolved, it is
displayed with a line through it.  Issues that are not viewable by the user due
to permissions do not have a line through them and do not have a summary shown.

Testing:
Ran a local instance of monorail and added multiple bugs with blocking and
duplicates issues.  Then view them as a list with the links below and also test
that clicking/middle-clicking on them works properly and that
Restrict-View-EditIssue issues do not leak info about summary nor open/closed
status while signed out.

http://localhost:8081/p/monorail-testing/issues/list?colspec=ID%20Type%20Status%20Priority%20Milestone%20Owner%20Summary%20BlockedOn%20Blocking%20MergedInto&q=&can=1
http://localhost:8081/u/2507136630/hotlists/test-list?can=1&q=&sort=&groupby=&colspec=Rank+Project+Status+Type+ID+Stars+Owner+Summary+Modified+Blocked+Blocking+BlockedOn

Bug: monorail:6280
Change-Id: I2e53bb3d2808c5913d4baf6895e565a836dd51fe
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1827980
Reviewed-by: Tiffany Zhang <zhangtiff@chromium.org>
Commit-Queue: Eyal Soha <eyalsoha@google.com>
Cr-Commit-Position: refs/heads/master@{#26117}